requirements.txt
langchain==0.1.15
langserve==0.0.51
langchain-openai==0.1.2
fastapi==0.110.1
sse-starlette==2.1.0
faiss-cpu==1.8.0
langchain-community==0.0.32
langchainhub==0.1.15
langchain-text-splitters==0.0.1
httpx-sse==0.4.0
httpx-sse==0.4.0
python-dotenv
langchain-huggingface==0.0.3
pypdf==4.3.1
rank_bm25==0.2.2
umap-learn==0.5.6
streamlit==1.37.1